+++ NEWS.md
@@ -1,8 +1,9 @@
 
 # BayesSPsurv 0.1.2
 
-* threshold argument added to `plot_Moran.I()` and `plot_JointCount()`
+* threshold argument added to `plot_Moran.I()` and `plot_JoinCount()`
 * parameter argument added to generic function `plot()`
+* rename plot_JoinCount(): `plot_JoinCount()`
 
 # BayesSPsurv 0.1.1